Create docker-image.yml
A Docker image is a lightweight, standalone, and executable package that contains everything needed to run a piece of software, including the code, runtime, system tools, libraries, and settings. It's essentially a snapshot of a specific application or service at a particular point in time. Docker images are used with Docker containers, which are instances created from these images